tried adding audio, added lanterns, did some other stuff, added weapon switching in a really bad way
Yo, what's up! I've been working on the game and made some progress. I tried adding audio to make it more immersive, but I gotta be real with you, it didn't work out. No worries though, I'll keep trying and see if I can get it to work in the future.

On a brighter note, I added some cool lanterns to the game that really set the vibe. I also did some other stuff, but I can't quite remember what it was. Oh yeah, I added weapon switching, but I gotta admit, it's not good. Like, it's really bad. Sorry about that, I'll make sure to fix it up real nice in the next update.

Despite the setbacks, I'm feeling pretty optimistic about the game. We'll keep on pushing forward and making it the best it can be. Yeah, science!